src/agent-skills-upd/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-skills-upd"
version = "0.0.2"
description = "Update your Claude Code, Codex, OpenCode, Amp, ClawdBot skills"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = ["httpx>=0.27", "typer>=0.12", "rich>=13.0", "pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1.0", "mypy>=1.0", "types-PyYAML>=6.0"]

[project.scripts]
skill-upd = "agent_skills_upd.cli.skill:app"
command-upd = "agent_skills_upd.cli.command:app"
agent-upd = "agent_skills_upd.cli.agent:app"
create-agent-skill-repo = "agent_skills_upd.cli.create:app"

[tool.hatch.build.targets.wheel]
packages = ["agent_skills_upd"]
